% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{host}
\alias{host}
\title{Host Function}
\usage{
host(model, domain = "hostr.so")
}
\arguments{
\item{model}{Your R model}

\item{domain}{Domain, where hoster cloud is hosted (defaults to hostr.so)}
}
\description{
This function uploads your model to the hostr cloud
}
\examples{
host(model)
host(model, "hostr.example.com")
}
\keyword{host}
